fix(metrics): update base_metric_computer imports after module rename
